Move Astropy code examples outside Example function.
To reduce clutter in the documentation, move the comment lines that give
the code to run in Python to get the AstroPy cosmology reference values
to outside the Example function.  It's important to list the reference
case in the code for the test, but it's not something the user needs
to know when looking up the Example of usage.