Use passed function name in runTestByName instead of runtime.
In preparation for full table-driven testing, where the test are run
by one function that's just stepping through the table, switch
runTestByName to quote the explicitly method name passed in
instead of using 'runtime.Caller' to quote where it was called from.

This also changes the function signature by removing the need for stackLevel.