Construct and call simpler cosmologies where possible.
WACDM->WCDM if WA=0
WCDM->LamdaCDM if W0=-1
LambdaCDM->FlatCDM if Om0+Ol0=1

Still direct special cases for *OM cosmologies
as W0 can be !=-1 if Ol0==0 and still satisfy OM.
Such a case is likely a mistake, but could arise from a parameter exploration.

This is instead of calling functions directly for special cases.